requirements.txt
streamlit==1.29.0
PyGithub==1.59.1
GitPython==3.1.40
streamlit-authenticator==0.2.3
requests==2.31.0
pathlib==1.0.1